+++ tmc/ic/TMC2226/TMC2226.c
@@ -68,7 +68,7 @@ int32_t tmc2226_readInt(TMC2226TypeDef *tmc2226, uint8_t address)
 	if (data[7] != tmc2226_CRC8(data, 7))
 		return 0;
 
-	return (data[3] << 24) | (data[4] << 16) | (data[5] << 8) | data[6];
+	return ((uint32_t)data[3] << 24) | ((uint32_t)data[4] << 16) | (data[5] << 8) | data[6];
 }
 
 void tmc2226_init(TMC2226TypeDef *tmc2226, uint8_t channel, uint8_t slaveAddress, ConfigurationTypeDef *tmc2226_config, const int32_t *registerResetState)
